requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.14
redis>=4.6.0
pytest>=7.4.0
httpx>=0.26.0
pytest-asyncio>=0.22.0
bandit>=1.7.0
python-dotenv>=1.0.0
pymysql>=1.0.0
pytest-cov>=4.1.0
flake8>=6.0.0
